pyproject.toml
[project]
name = "hacknation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "google-cloud-aiplatform[adk,agent-engines]>=1.128.0",
    "pydantic-ai>=1.27.0",
    "pydantic-settings>=2.12.0",
    "google-generativeai>=0.8.0",
    "python-dotenv>=1.2.1",
    "pydantic-ai-slim[google,logfire]>=1.27.0",
    "python-dotenv>=1.2.1",
    "logfire>=4.15.1",
    "sentence-transformers>=5.1.2",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.13.0",
    "pgvector>=0.3.0",
    "sqlalchemy>=2.0.0",
    "psycopg2>=2.9.11",
    "pypdf>=4.0.0",
    "transformers>=4.30.0",
    "torch>=2.0.0",
    # "sentence-transformers>=2.2.0",
    # "psycopg2-binary>=2.9.0",
    # "pgvector>=0.2.0",
    "asyncpg>=0.31.0",
    "greenlet>=3.3.0",
]
